Remove null values and fix visualization script
Remove all observations with null values except for records only
containing null values in the "nonwhite" and "south" columns in the
year 2021. This is because the GSS was conducted online in 2021 and
data for those variables were not collected.

Create two separate visualizations for the data set including and
excluding null values in columns other than those plotted.