Fix data transformation issues
Fix issue where the conservative dummy was not derived properly. This
issue caused individuals identifying as "extremely conservative" and
"slightly conservative" to be classified as "liberal". This problem
affected the "gss" and the "gss-with-na" data files.

Additionally, this commit includes individuals of age "89 or older", a
group previously omitted from the sample.

The data structure of the "gss-raw" data set changes to reflect the
initial data structures when loading the file.

The above-mentioned issues affected the visualizations of trust in
science by political ideology ("consci-by-polviews") which this commit
fixes.